requirements.txt
alembic==1.8.0
autopep8==1.6.0
bcrypt==3.2.2
blinker==1.6.2
cffi==1.15.0
click==8.1.6
colorama==0.4.4
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==4.11.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.34
toml==0.10.2
Werkzeug==2.3.6
zipp==3.8.0
